tests/wpt/tests/html/semantics/embedded-content/media-elements/interfaces/HTMLElement/HTMLMediaElement/textTracks.html

<!doctype html>
<title>HTMLMediaElement.textTracks</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
var video = document.createElement('video');
test(function(){
    assert_equals(video.textTracks, video.textTracks);
    assert_equals(video.textTracks.length, 0);
});

</script>
